+++ src/main/java/com/library/model/Admin.java
@@ -1,0 +1,66 @@
+package com.library.model;
+
+public class Admin {
+    private int adminId;
+    private String username;
+    private String password;
+    private String name;
+
+    // Constructor for creating a new Admin instance
+    public Admin(String username, String password, String name) {
+        this.username = username;
+        this.password = password;
+        this.name = name;
+    }
+
+    // Constructor with adminId for fetching existing Admin from the database
+    public Admin(int adminId, String username, String password, String name) {
+        this.adminId = adminId;
+        this.username = username;
+        this.password = password;
+        this.name = name;
+    }
+
+    // Getters and Setters
+    public int getAdminId() {
+        return adminId;
+    }
+
+    public void setAdminId(int adminId) {
+        this.adminId = adminId;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String toString() {
+        return "Admin{" +
+                "adminId=" + adminId +
+                ", username='" + username + '\'' +
+                ", password='" + password + '\'' +
+                ", name='" + name + '\'' +
+                '}';
+    }
+}
