--- src/main/java/model/Admin.java
@@ -1,26 +1,0 @@
-// src/main/java/model/Admin.java
-package model;
-
-public class Admin {
-    private String adminId;
-    private String name;
-
-    public Admin() {
-    }
-
-    public String getAdminId() {
-        return adminId;
-    }
-
-    public void setAdminId(String adminId) {
-        this.adminId = adminId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
