--- src/main/java/model/Librarian.java
@@ -1,44 +1,0 @@
-// src/main/java/model/Librarian.java
-package model;
-
-public class Librarian {
-    private String librarianId;
-    private String name;
-    private String section;
-    private boolean isApproved;
-
-    public Librarian() {
-    }
-
-    public String getLibrarianId() {
-        return librarianId;
-    }
-
-    public void setLibrarianId(String librarianId) {
-        this.librarianId = librarianId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSection() {
-        return section;
-    }
-
-    public void setSection(String section) {
-        this.section = section;
-    }
-
-    public boolean isApproved() {
-        return isApproved;
-    }
-
-    public void setApproved(boolean approved) {
-        isApproved = approved;
-    }
-}
